fix: Apply the correct types for data fields
- Field "weight": Originally String (Ex: "45kg") to represent both the
weight and its measurement unit. This leads to parsing issue and
future compatibility conflicts since the measurement unit may change.
Has been changed to Double

- Field "kilometers": Originally Long which is not true in reality (You
can still have 0.2 km or 0.5 km). Has been changed to Double

+++ app/src/main/java/com/haidoan/android/stren/core/datasource/model/FirestoreTrainedExercise.kt
@@ -25,7 +25,7 @@ internal data class FirestoreTrainedExercise(
                 firestoreTrainingSets.forEach {
                     trainingSets.add(
                         TrainingMeasurementMetrics.DistanceAndDuration(
-                            it.key.toLong(),
+                            it.key.toDouble(),
                             it.value.toDouble()
                         )
                     )
@@ -38,7 +38,12 @@ internal data class FirestoreTrainedExercise(
             }
             else -> {
                 firestoreTrainingSets.forEach {
-                    trainingSets.add(TrainingMeasurementMetrics.WeightAndRep(it.key, it.value))
+                    trainingSets.add(
+                        TrainingMeasurementMetrics.WeightAndRep(
+                            it.key.toDouble(),
+                            it.value
+                        )
+                    )
                 }
             }
         }
